requirements.txt
annotated-types==0.7.0
asgiref==3.9.2
crispy-bootstrap5==2025.6
Django==5.2.6
django-crispy-forms==2.4
django-filter==25.1
django-js-asset==3.1.2
django-mptt==0.17.0
django-qr-code==4.2.0
django-qrcode==0.3
django-tables2==2.7.0
djangorestframework==3.16.1
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
pytz==2025.2
PyYAML==6.0.2
segno==1.6.6
sqlparse==0.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
Pillow
psycopg2
django-environ==0.12.0
